valkyire/lifelink_backend/lifelink_backend/requirements.txt
Flask
Flask-CORS
mysql-connector-python
anthropic
twilio